+++ semver-gradle-plugin/main/kotlin/SemVerExtension.kt
@@ -7,10 +7,8 @@ import org.gradle.kotlin.dsl.getByType
 public interface SemVerExtension {
 
     public val tagPrefix: Property<String>
-    public val applyToAllProjects: Property<Boolean>
 
     public companion object {
-        internal const val defaultApplyToAllProjects = true
         internal const val defaultTagPrefix = ""
         internal const val name = "semver"
     }
@@ -21,6 +19,3 @@ internal val Project.semVerExtension: SemVerExtension
 
 internal val Project.tagPrefix: String
     get() = semVerExtension.tagPrefix.get()
-
-internal val Project.applyToAllProjects: Boolean
-    get() = semVerExtension.applyToAllProjects.get()