+++ semver-gradle-plugin/test/kotlin/VersionBuildDirTest.kt
@@ -1,6 +1,7 @@
 package com.javiersc.semver.gradle.plugin
 
 import com.javiersc.semver.gradle.plugin.internal.calculateAdditionalVersionData
+import com.javiersc.semver.gradle.plugin.internal.headRevCommitInBranch
 import io.kotest.matchers.shouldBe
 import java.io.File
 import java.time.Instant
@@ -11,31 +12,41 @@ import org.gradle.testkit.runner.BuildResult
 class VersionBuildDirTest {
 
     @Test
-    fun `clean without tag in current commit`() =
+    fun `clean without tag in current commit - hash`() =
         testSandbox(
-            sandboxPath = "semver/version-build-dir/clean-with-no-tag",
+            sandboxPath = "semver/version-build-dir/clean-with-no-tag (hash)",
             beforeTest = {
                 generateInitialCommitAddVersionTagAndAddNewCommit()
                 File("$this/expect-version.txt").apply {
                     createNewFile()
-                    writeText("1.0.0${parentFile.git.calculateAdditionalVersionData()}\n")
+                    writeText(expectVersion)
                 }
             },
             test = ::testSemVerVersionBuildDir,
         )
 
     @Test
-    fun `no clean without tag in current commit`() =
+    fun `no clean without tag in current commit - timestamp`() =
         testSandbox(
-            sandboxPath = "semver/version-build-dir/no-clean-with-no-tag",
+            sandboxPath = "semver/version-build-dir/no-clean-with-no-tag (timestamp)",
             beforeTest = {
-                val mockDate = Date.from(Instant.ofEpochSecond(0))
                 generateInitialCommitAddVersionTagAndAddNewCommit()
                 File("$this/new-2.txt").createNewFile()
                 File("$this/expect-version.txt").apply {
                     createNewFile()
-                    writeText("1.0.0${parentFile.git.calculateAdditionalVersionData(mockDate)}\n")
+                    writeText(expectVersion)
                 }
+            },
+            test = ::testSemVerVersionBuildDir,
+        )
+
+    @Test
+    fun `clean 1_0_0`() =
+        testSandbox(
+            sandboxPath = "semver/version-build-dir/clean 1_0_0",
+            beforeTest = {
+                generateInitialCommitAddVersionTagAndAddNewCommit()
+                git.tag().setObjectId(git.headRevCommitInBranch).setName("v1.0.0").call()
             },
             test = ::testSemVerVersionBuildDir,
         )
@@ -44,6 +55,17 @@ class VersionBuildDirTest {
 @Suppress("UNUSED_PARAMETER")
 internal fun testSemVerVersionBuildDir(result: BuildResult, testProjectDir: File) {
     val version = File("$testProjectDir/build/semver/version.txt").readText()
-    val expectVersion = File("$testProjectDir/expect-version.txt").readLines().first()
+    val expectVersion = File("$testProjectDir/expect-version.txt").readText()
     version shouldBe expectVersion
 }
+
+internal val File.expectVersion: String
+    get() {
+        val mockDate = Date.from(Instant.ofEpochSecond(0))
+        val additionalData = parentFile.git.calculateAdditionalVersionData(mockDate)
+        return """
+            |1.0.0$additionalData
+            |v1.0.0$additionalData
+            |
+        """.trimMargin()
+    }