+++ CHANGELOG.md
@@ -13,9 +13,9 @@
 ### Fixed
 
 ### Updated
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.8`
 - `gradle -> 7.3`
 - `org.jetbrains.kotlinx:binary-compatibility-validator -> 0.8.0`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.4`
 
 
 ## [0.0.0] - 2021-10-15